% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stm_transform_insert.R
\name{stm_insert}
\alias{stm_insert}
\alias{stm_insert_vector}
\alias{stm_insert_rows}
\alias{stm_insert_cols}
\title{Insert values into a matrix or vector}
\usage{
stm_insert(dat, insert_at = length(dat), values, axis = NULL)

stm_insert_vector(vec, insert_at = length(vec), values)

stm_insert_rows(mat, insert_at = length(mat), values)

stm_insert_cols(mat, insert_at = length(mat), values)
}
\arguments{
\item{dat}{The matrix, vector, or array to which new value(s) will be added.
\describe{
  \item{In \code{stm_insert}}{Can be a vector or matrix.}
  \item{In \code{stm_insert_vector}}{Specifically a vector.}
  \item{In \code{stm_insert_rows}}{Specifically a matrix.}
  \item{In \code{stm_insert_cols}}{Specifically a matrix.}
}}

\item{insert_at}{Optional. The location where to insert the new value(s). 
If \code{length(dat)}, value(s) are inserted at the 
end of the matrix or vector.}

\item{values}{The value(s) to insert as new columns or rows.}

\item{axis}{Optional. The axis along which to insert the values: 
\code{0} for rows, \code{1} for columns. Default is \code{NULL},
meaning the values will be inserted into the vector.}

\item{vec}{A vector of values.}

\item{mat}{A matrix of values.}
}
\value{
A matrix or vector with the new value(s) inserted at the specified
        location.
}
\description{
\code{stm_insert()} and its variants insert value(s) into a matrix or
vector at a specified location. The function allows inserting data
into either rows or columns, or simply appending values to the end.
}
\examples{
v <- 11:22
print(v)
stm_insert(v, insert_at = 3, values = 101:103)

A <- matrix(dat = v, nrow = 3, byrow = TRUE)
print(A)
stm_insert(A, insert_at = 1, values = 101:104, axis = 0)

A <- matrix(dat = 11:22, nrow = 3, byrow = TRUE)
print(A)
stm_insert(A, insert_at = 3, values = 101:103, axis = 1)

stm_insert_vector(11:22, insert_at = 1, values = 101:104)
A <- matrix(dat = 11:22, nrow = 3, byrow = TRUE)
stm_insert_rows(A, insert_at = 1, values = 101:104)

A <- matrix(dat = 11:22, nrow = 3, byrow = TRUE)
stm_insert_cols(A, insert_at = 3, values = 101:103)
}
