% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stm_constants.R
\docType{data}
\name{stm_12th_root_of_root2}
\alias{stm_12th_root_of_root2}
\title{Twelfth root of root 2}
\format{
A numeric value.
}
\usage{
stm_12th_root_of_root2
}
\description{
A predefined constant for the 12th root of root 2: \code{sqrt(2)^(1/12)}.
}
\keyword{datasets}
