% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stm_constants.R
\docType{data}
\name{stm_2_root_2}
\alias{stm_2_root_2}
\title{2 raised to root 2}
\format{
A numeric value.
}
\usage{
stm_2_root_2
}
\description{
A predefined constant for 2 raised to the power of sqrt(2): \code{2^sqrt(2)}.
}
\keyword{datasets}
