% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stm_companion.R
\name{stm_companion}
\alias{stm_companion}
\title{Compute the companion matrix from a given vector}
\usage{
stm_companion(vec)
}
\arguments{
\item{vec}{A numeric vector. The first element represents the coefficient 
of the highest degree term.}
}
\value{
A companion matrix constructed from the input vector.
}
\description{
\code{stm_companion()} computes the companion matrix from a given vector 
by generating the matrix form for a linear recurrence relation.
}
\details{
The companion matrix is created from the vector by setting the first row 
to the negative of the elements (except the first) divided by the first 
element, and setting subdiagonal elements to 1.
}
\examples{
v <- c(2, 8, 5, 10, 18)
print(v)
stm_companion(v)

}
