% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stm_constant.R
\name{stm_constant}
\alias{stm_constant}
\title{Create a constant matrix}
\usage{
stm_constant(k = NULL, nrows = 1, ncols = nrows, digits = NULL)
}
\arguments{
\item{k}{The constant value to fill the matrix. If \code{NULL}, a random 
value is selected from the range -5 to 9.}

\item{nrows}{The number of rows in the matrix. Must be at least 1.}

\item{ncols}{The number of columns in the matrix. Must be at least 1.}

\item{digits}{Optional. The number of decimal places to round the matrix 
values. Default is \code{NULL}.}
}
\value{
A matrix filled with the constant value \code{k}.
}
\description{
\code{stm_constant()} creates a matrix of a specified size where all 
elements are the same constant value. If no constant value is provided, 
one is randomly selected from the range -5 to 9.
}
\details{
This function creates a matrix of the specified size, and all elements 
are filled with the constant value \code{k}. If no value is specified, 
one is randomly selected from the range -5 to 9.
}
\examples{
stm_constant(k = 3, nrows = 4)
stm_constant(nrows = 3, ncols = 5)

}
