% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stm_arrays_like.R
\name{stm_constant_like}
\alias{stm_constant_like}
\title{Generate a constant matrix or vector like an existing matrix or vector}
\usage{
stm_constant_like(vmat, k = 2)
}
\arguments{
\item{vmat}{A matrix or vector whose dimensions will be used to generate
the constant matrix or vector.}

\item{k}{A numeric value representing the constant to fill the generated
matrix or vector. Default is \code{2}.}
}
\value{
A matrix or vector of the same dimensions as \code{vmat}, filled
        with the constant value \code{k}.
}
\description{
\code{stm_constant_like()} generates a constant matrix or vector that
matches the dimensions of the provided input matrix or vector.
}
\details{
The function checks the validity of the input \code{vmat} and \code{k},
converting \code{vmat} to a matrix if it is a vector. It then generates
a matrix of the same size as \code{vmat} filled with the constant value
\code{k}.
}
\examples{
A <- matrix(1:12, nrow = 3, byrow = TRUE)
print(A)
stm_constant_like(A, k = 5)
stm_constant_like(c(1, 2, 3), k = 7)

}
\seealso{
\code{\link{stm_constant}}
}
