% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stm_constants.R
\docType{data}
\name{stm_cuberoot_of_root2}
\alias{stm_cuberoot_of_root2}
\title{Cube root of root 2}
\format{
A numeric value.
}
\usage{
stm_cuberoot_of_root2
}
\description{
A predefined constant for the cube root of root 2: \code{sqrt(2)^(1/3)}.
}
\keyword{datasets}
