% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stm_constants.R
\docType{data}
\name{stm_cuberoot_of_root3}
\alias{stm_cuberoot_of_root3}
\title{Cube root of root 3}
\format{
A numeric value.
}
\usage{
stm_cuberoot_of_root3
}
\description{
A predefined constant for the cube root of root 3: \code{sqrt(3)^(1/3)}.
}
\keyword{datasets}
