% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stm_trigs.R
\name{stm_deg_to_radians}
\alias{stm_deg_to_radians}
\title{Convert degrees to radians}
\usage{
stm_deg_to_radians(degrees, digits = NULL)
}
\arguments{
\item{degrees}{A numeric value representing the angle in degrees.}

\item{digits}{Optional. The number of decimal places to round the result.
Default is \code{NULL} for no rounding.}
}
\value{
The value in radians, rounded to the specified number of decimal 
        places if \code{digits} is provided.
}
\description{
\code{stm_deg_to_radians()} converts a value in degrees to radians. 
The result can be rounded to a specified number of decimal places.
}
\details{
The function converts the given angle in degrees to radians by multiplying 
by \code{pi / 180}. If \code{digits} is specified, the result is rounded 
to that number of decimal places.
}
\examples{
stm_deg_to_radians(90)
stm_deg_to_radians(c(0, 45, 90))
stm_deg_to_radians(180, digits = 2)

}
\seealso{
\code{\link{stm_rad_to_degrees}}
}
