% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stm_tranform_keep_delete.R
\name{stm_delete_cols}
\alias{stm_delete_cols}
\title{Delete specified columns from a data frame}
\usage{
stm_delete_cols(dat, indices, as_matrix = TRUE)
}
\arguments{
\item{dat}{A data frame or matrix from which columns will be removed.}

\item{indices}{A numeric vector or logical vector indicating the columns
to delete.}

\item{as_matrix}{Optional. A logical value indicating whether the result 
should be returned as a matrix. Default is \code{TRUE}.}
}
\value{
A data frame or matrix with the specified columns removed.
}
\description{
\code{stm_delete_cols()} removes specified columns from a data frame.
It uses \code{stm_transform_keep_delete()} to perform the operation.
}
\details{
This function uses \code{stm_transform_keep_delete()} to remove the specified
columns from the given data. If \code{as_matrix} is \code{TRUE}, the result
is returned as a matrix, otherwise a data frame is returned.
}
\examples{
A <- matrix(1:15, ncol = 5, byrow = TRUE)
print(A)
stm_delete_cols(A, 2)

}
\seealso{
\code{\link{stm_keep_cols}}, \code{\link{stm_delete_rows}}, \code{\link{stm_keep_rows}}
}
