% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stm_digits.R
\name{stm_digits}
\alias{stm_digits}
\title{Set default number of digits for computation}
\usage{
stm_digits(digits = 14)
}
\arguments{
\item{digits}{An integer specifying the number of decimal digits. Default 
is 14.}
}
\value{
An integer representing the number of decimal digits to use.
}
\description{
\code{stm_digits()} determines the number of decimal digits to be used 
for computations, providing a default value if no valid input is 
supplied.
}
\details{
This function checks if the provided \code{digits} parameter is a valid 
integer. If the input is invalid or \code{NULL}, the function defaults 
to 14 decimal digits.
}
\examples{
stm_digits(10)

stm_digits(NULL)

stm_digits("abc")

stm_digits()

}
