% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stm_constants.R
\docType{data}
\name{stm_euler_gamma}
\alias{stm_euler_gamma}
\title{Euler-Gamma constant}
\format{
A numeric value.
}
\usage{
stm_euler_gamma
}
\description{
A predefined constant for the Euler-Gamma constant: \code{euler_gamma}.
}
\keyword{datasets}
