% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stm_constants.R
\docType{data}
\name{stm_euler_gompertz_delta}
\alias{stm_euler_gompertz_delta}
\title{Euler-Gompertz constant}
\format{
A numeric value.
}
\usage{
stm_euler_gompertz_delta
}
\description{
A predefined constant for the Euler-Gompertz constant: \code{0.59634736232319407434}.
}
\keyword{datasets}
