% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stm_constants.R
\docType{data}
\name{stm_exp}
\alias{stm_exp}
\title{Base of natural logarithms}
\format{
A numeric value.
}
\usage{
stm_exp
}
\description{
A predefined constant for the base of natural logarithms: \code{exp(1)}.
}
\keyword{datasets}
