% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stm_constants.R
\docType{data}
\name{stm_exp_pi}
\alias{stm_exp_pi}
\title{Exponential of pi}
\format{
A numeric value.
}
\usage{
stm_exp_pi
}
\description{
A predefined constant for e raised to the power of pi: \code{exp(pi)}.
}
\keyword{datasets}
