% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stm_eye.R
\name{stm_eye}
\alias{stm_eye}
\title{Create a matrix with ones on the diagonal and zeros elsewhere}
\usage{
stm_eye(nrows, ncols = nrows, k = 0)
}
\arguments{
\item{nrows}{An integer specifying the number of rows in the matrix. Must 
be greater than or equal to zero.}

\item{ncols}{An integer specifying the number of columns in the matrix. 
Defaults to \code{nrows}. Must be greater than or equal to zero.}

\item{k}{Optional. An integer specifying the diagonal offset. The default 
is \code{0} for the main diagonal. A positive value retrieves 
diagonals above the main diagonal, while a negative value 
retrieves diagonals below the main diagonal.}
}
\value{
A matrix of dimensions \code{nrows} x \code{ncols} with ones along 
        the diagonal (or at the offset specified by \code{k}), and zeros 
        elsewhere.
}
\description{
\code{stm_eye()} creates a matrix with ones along the diagonal, allowing 
for an offset of the diagonal (parameter \code{k}).
}
\details{
The function generates a matrix where ones appear on the diagonal (or an 
offset diagonal determined by \code{k}), and zeros appear everywhere else. 

If \code{k} exceeds the valid range (i.e., the diagonal index is out of 
bounds), a matrix of zeros is returned.
}
\examples{
stm_eye(nrows = 3)
stm_eye(nrows = 4, k = 1)
stm_eye(nrows = 4, ncols = 5, k = -2)

}
\seealso{
\code{\link{stm_identity}}
}
