% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stm_tranform_fliprotate.R
\name{stm_fliplr}
\alias{stm_fliplr}
\title{Flip a matrix left to right}
\usage{
stm_fliplr(mat)
}
\arguments{
\item{mat}{A numeric matrix to be flipped left to right.}
}
\value{
A matrix that is flipped left to right.
}
\description{
\code{stm_fliplr()} flips a matrix left to right, i.e., it reverses the 
order of columns. This is achieved by calling the \code{stm_flip()} function 
with \code{left_right = TRUE}.
}
\details{
This function uses the \code{stm_flip()} function with the argument 
         \code{left_right = TRUE} to flip the matrix horizontally.
}
\examples{
A <- matrix(1:12, ncol = 4, byrow = TRUE)
print(A)
stm_fliplr(A)

}
\seealso{
\code{\link{stm_flipud}}
}
