% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stm_tranform_fliprotate.R
\name{stm_flipud}
\alias{stm_flipud}
\title{Flip a matrix upside down}
\usage{
stm_flipud(mat)
}
\arguments{
\item{mat}{A numeric matrix to be flipped upside down.}
}
\value{
A matrix that is flipped upside down.
}
\description{
\code{stm_flipud()} flips a matrix upside down, i.e., it reverses the 
order of rows. This is achieved by calling the \code{stm_flip()} function 
with \code{left_right = FALSE}.
}
\details{
This function uses the \code{stm_flip()} function with the argument 
         \code{left_right = FALSE} to flip the matrix vertically.
}
\examples{
A <- matrix(1:16, ncol = 4, byrow = TRUE)
print(A)
stm_flipud(A)

}
\seealso{
\code{\link{stm_fliplr}}
}
