% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stm_fsystem.R
\name{stm_fsystem}
\alias{stm_fsystem}
\title{Create and evaluate a system of expressions with vector inputs}
\usage{
stm_fsystem(fsystem, vars = NULL)
}
\arguments{
\item{fsystem}{A character vector where each element is a mathematical
expression as a string, representing the equations.}

\item{vars}{Optional. A character vector specifying the variable names
used in the expressions. Default is \code{NULL}, meaning
that variables are automatically detected.}
}
\value{
A function that takes vectors of arguments and evaluates each
        expression in \code{fsystem} with the given values, returning
        a numeric vector of results.
}
\description{
\code{stm_fsystem()} converts a character vector of expressions into
a function that evaluates each expression using vectors of provided
arguments.
}
\details{
This function takes a system of mathematical expressions and converts
them into a callable R function. Each expression is evaluated with
corresponding input vectors. If the number of provided arguments does
not match the number of expressions, an error is raised.
}
\examples{
f <- c("2 * x + y", "x^2 + sqrt(y)")
f_list <- stm_fsystem(f, vars = c("x", "y"))
f_list(c(x = 4, y = 9), c(x = 12, y = 7))

f <- c("a + b", "a^2 - b")
f_list <- stm_fsystem(f, vars = c("a", "b"))
f_list(c(a = 3, b = 2), c(a = 5, b = 1))

}
\seealso{
\code{\link{stm_function}}, \code{\link{stm_fexpr_vars}}
}
