% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stm_function.R
\name{stm_function}
\alias{stm_function}
\title{Convert string expression to function}
\usage{
stm_function(fexpr, vars = NULL, nvars = NULL, par_name = "fexpr")
}
\arguments{
\item{fexpr}{A function or string representing a mathematical expression. 
Strings must represent valid mathematical expressions.}

\item{vars}{Optional. A character vector of expected variable names.
Default is \code{NULL} meaning that variables will be assumed
for multi-variable expressions.}

\item{nvars}{Optional. An integer specifying the expected number of 
variables in the expression. Raises an error if the count
differs. Default is \code{NULL}}

\item{par_name}{Optional. A string used in error and warning messages to 
refer to the parameter. Defaults is \code{"fexpr"}.}
}
\value{
If \code{fexpr} is a function, it is returned unchanged. If it is
        a string, a callable R function is constructed and returned.
}
\description{
\code{stm_function()} converts a function or string representation of an
expression into a valid R function. Checks variable compatibility and
provides warnings or errors as necessary.
}
\details{
The function checks whether \code{fexpr} is a valid function or expression.
If it is a string, variables are identified and a callable function is
constructed. If \code{nvars} is specified and does not match the variable
count, an error is raised. If \code{vars} is not specified and multiple
variables are found, a warning is issued and default names are assumed.
}
\examples{
f <- stm_function(
  fexpr = "2 * x + sqrt(y) + cos(a) / pi", 
  vars = c("x", "y", "a")
)
f(3, 4, 5)
f(x = 3, y = 4, a = 5)
f(a = 3, x = 4, y = 5)

}
\seealso{
\code{\link{stm_function}}, \code{\link{stm_fexpr_vars}}
}
