% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stm_constants.R
\docType{data}
\name{stm_gauss}
\alias{stm_gauss}
\title{Gauss constant}
\format{
A numeric value.
}
\usage{
stm_gauss
}
\description{
A predefined constant for the Gauss constant: \code{(gamma(1/4)^2) / (2 * sqrt(2 * pi^3))}.
}
\keyword{datasets}
