% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stm_gtable.R
\name{stm_gtable}
\alias{stm_gtable}
\title{Create a styled table with gt}
\usage{
stm_gtable(dat)
}
\arguments{
\item{dat}{A data frame or tibble to be converted into a styled table. 
It will be converted to a tibble if necessary.}
}
\value{
A `gt` table object with styling applied. The table will have missing 
        values replaced with a dash ("-"), cell borders, and left alignment 
        for character columns.
}
\description{
\code{stm_gtable()} converts the provided data to a tibble, replaces any 
missing values with a dash ("-"), and generates a styled table using the 
\code{gt} package. It applies borders, alignment, and other custom styles 
to the table's body and column labels.
}
\details{
The function performs the following operations:
\itemize{
  \item Converts the input data into a tibble using \code{tibble::as_tibble()}.
  \item Replaces any missing values in the data with a dash ("-") using 
        \code{tidyr::replace_na()}.
  \item Creates a table using the \code{gt::gt()} function.
  \item Applies cell border styles to the body and column labels of the table.
  \item Left-aligns character columns and applies additional styling such as 
        padding and bold column labels.
}
}
\examples{
tbl <- tibble::tibble(
  Name = c("John", "Jane", "Doe"),
  Age = c(25, 30, NA),
  City = c("New York", "Los Angeles", "Chicago")
)
stm_gtable(tbl)

}
\seealso{
\code{\link[gt]{gt}}, \code{\link[tidyr]{replace_na}}, \code{\link[tibble]{as_tibble}}
}
