% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stm_hadamard.R
\name{stm_hadamard}
\alias{stm_hadamard}
\title{Generate Hadamard matrix}
\usage{
stm_hadamard(n)
}
\arguments{
\item{n}{An integer specifying the size of the matrix. It must be a
power of 2.}
}
\value{
A Hadamard matrix of size n x n.
}
\description{
\code{stm_hadamard()} generates a Hadamard matrix of size n, where n
must be a power of 2. The matrix is recursively built by combining
smaller Hadamard matrices.
}
\details{
The function checks whether n is a valid power of 2. If n is not a power 
of 2, an error is raised. The function starts with a 1x1 matrix and 
recursively expands it by creating larger matrices by joining smaller ones.
}
\examples{
stm_hadamard(4)
stm_hadamard(8)
stm_hadamard(16)

}
