% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stm_hankel.R
\name{stm_hankel}
\alias{stm_hankel}
\title{Create Hankel matrix}
\usage{
stm_hankel(vec1, vec2 = NULL)
}
\arguments{
\item{vec1}{A numeric vector to form the first column of the Hankel matrix.}

\item{vec2}{Optional. A numeric vector to form the first row of the Hankel
matrix. If not provided, the last element of \code{vec1} is 
used to initialize the first element of \code{vec2}.}
}
\value{
A Hankel matrix of size \code{m x n}, where \code{m} is the length 
        of \code{vec1} and \code{n} is the length of \code{vec2}.
}
\description{
\code{stm_hankel()} generates a Hankel matrix from two vectors, where
the first vector forms the first column and the second vector forms the
first row of the matrix. If the second vector is not provided, the last
element of the first vector is used to initialize the first element of
the second vector.
}
\details{
The function constructs a Hankel matrix where the first column is formed 
from \code{vec1} and the first row is formed from \code{vec2}. If \code{vec2} 
is not provided, the last element of \code{vec1} is used as the first element 
of \code{vec2}. The matrix is populated such that each element in the 
matrix is determined by the sum of its row and column index.
}
\examples{
x <- c(1, 2, 3, 5)
y <- c(5, 6, 7)
stm_hankel(x, y)

}
