% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stm_hilbert.R
\name{stm_hilbert}
\alias{stm_hilbert}
\title{Generate Hilbert matrix}
\usage{
stm_hilbert(n)
}
\arguments{
\item{n}{A positive integer specifying the size of the Hilbert matrix. 
It represents both the number of rows and columns.}
}
\value{
An \code{n x n} matrix with elements \code{Hij = 1 / (i + j - 1)}.
}
\description{
\code{stm_hilbert()} creates an \code{n x n} Hilbert matrix, where each 
element is defined as \code{Hij = 1 / (i + j - 1)}, where \code{i} and 
\code{j} are the row and column indices.
}
\details{
The Hilbert matrix is used in numerical analysis and is known for 
its ill-conditioning as n increases. This function uses the
\code{outer()} function to generate the matrix efficiently.
}
\examples{
stm_hilbert(3)
stm_hilbert(5)

}
