% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stm_identity.R
\name{stm_identity}
\alias{stm_identity}
\title{Generate identity matrix}
\usage{
stm_identity(n)
}
\arguments{
\item{n}{A positive integer specifying the size of the identity matrix. 
It represents both the number of rows and columns.}
}
\value{
An \code{n x n} identity matrix.
}
\description{
\code{stm_identity()} creates an \code{n x n} identity matrix where all 
diagonal elements are 1, and all off-diagonal elements are \code{0}.
}
\details{
The identity matrix is a square matrix with ones on the diagonal 
and zeros elsewhere. It is used in various mathematical operations such 
as matrix inversion and solving systems of linear equations.
}
\examples{
stm_identity(3)
stm_identity(5)

}
\seealso{
\code{\link{diag}}, \code{\link{stm_eye}}
}
