% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stm_ivps_funcs.R
\name{stm_ivps_meuler}
\alias{stm_ivps_meuler}
\title{Solve initial value problems (IVPs) using numerical methods}
\usage{
stm_ivps_meuler(
  odeqtn,
  exactsol = NULL,
  vars = c("t", "y"),
  time_span = NULL,
  y0 = NULL,
  step_size = NULL,
  nsteps = 10,
  maxit = 100,
  show_iters = NULL,
  view_table = FALSE,
  show_plot = TRUE,
  digits = 8
)
}
\arguments{
\item{odeqtn}{A function defining the ODE system. It should take two 
arguments: the current time and state vector as 
\code{y = f(t, y)}.}

\item{exactsol}{Optional. A function representing the exact solution for
comparison. Default is \code{NULL}.}

\item{vars}{Independent and dependent variables.}

\item{time_span}{A numeric vector specifying the time range for the
solution. Should be of length \code{2}.}

\item{y0}{Initial state values for the ODE system.}

\item{step_size}{The step size for the numerical method. If \code{NULL},
it is computed based on the time span and number of steps.}

\item{nsteps}{The number of time steps to take. Default is \code{10}.}

\item{maxit}{Maximum number of iterations for iterative methods.
Default is \code{10}.}

\item{show_iters}{Logical. If \code{TRUE}, iteration details are displayed.
Default is \code{NULL}.}

\item{view_table}{Logical. If \code{TRUE}, the results are viewed in the 
R spreadsheet. Default is \code{FALSE}.}

\item{show_plot}{Logical. If \code{TRUE}, the results are plotted. Default 
is \code{TRUE}.}

\item{digits}{The number of significant digits to display in the results.
Default is \code{8}.}
}
\value{
A data frame containing the time points and corresponding solution
        values. If an exact solution is provided, the error at each time
        point is also included.
\itemize{
    \item \strong{Time (t)}: Time points.
    \item \strong{Approx. (yi)}: State values at each time point.
    \item \strong{Exact (y)}: Exact solution if \code{exactsol} is provided.
    \item \strong{Error: |y - yi|}: The error at each time point if 
          \code{exactsol} is provided.
}
}
\description{
\code{stm_ivps*()} set of functions solve IVPs for ordinary differential 
equations (ODEs) using various numerical methods. Supports solving 
single equations or systems of ODEs.
}
\details{
These functions solve initial value problems for ODEs using a specified
numerical method. They can handle both single ODEs and systems of ODEs by
providing state variables as a vector. If an exact solution is provided,
the error at each time point will be calculated for comparison.
}
\examples{
f <- "y - t^2 + 1"
ft <- "(t + 1)^2 - 0.5 * exp(t)"
result <- stm_ivps_meuler(
  odeqtn = f,
  exactsol = ft,
  time_span = c(0, 2),
  y0 = 0.5,
  nsteps = 10,
  show_plot = FALSE
)
stm_gtable(result)

}
