% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stm_tranform_keep_delete.R
\name{stm_keep_cols}
\alias{stm_keep_cols}
\title{Keep specified columns in a data frame}
\usage{
stm_keep_cols(dat, indices, as_matrix = TRUE)
}
\arguments{
\item{dat}{A data frame or matrix from which columns will be retained.}

\item{indices}{A numeric vector or logical vector indicating the columns
to retain.}

\item{as_matrix}{Optional. A logical value indicating whether the result 
should be returned as a matrix. Default is \code{TRUE}.}
}
\value{
A data frame or matrix containing only the specified columns.
}
\description{
\code{stm_keep_cols()} retains specified columns from a data frame.
It uses \code{stm_transform_keep_delete()} to perform the operation.
}
\details{
This function uses \code{stm_transform_keep_delete()} to retain the specified
columns in the given data. If \code{as_matrix} is \code{TRUE}, the result
is returned as a matrix, otherwise a data frame is returned.
}
\examples{
A <- matrix(1:15, ncol = 5, byrow = TRUE)
print(A)
stm_keep_cols(A, 2)
stm_keep_cols(A, 2, as_matrix = FALSE)

}
\seealso{
\code{\link{stm_keep_rows}}, \code{\link{stm_delete_cols}}, \code{\link{stm_keep_rows}}
}
