% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stm_tranform_keep_delete.R
\name{stm_keep_rows}
\alias{stm_keep_rows}
\title{Keep specific rows in a data frame or matrix}
\usage{
stm_keep_rows(dat, indices, as_matrix = TRUE)
}
\arguments{
\item{dat}{A data frame or matrix from which rows are to be kept.}

\item{indices}{A vector of row indices to retain.}

\item{as_matrix}{A logical value indicating whether the result should 
be converted to a matrix. Default is \code{TRUE}.}
}
\value{
A data frame or matrix with the specified rows retained.
}
\description{
\code{stm_keep_rows()} keeps the specified rows from a data frame or 
matrix based on the provided indices.
}
\details{
This function uses \code{stm_transform_keep_delete()} to retain the rows 
specified in \code{indices}. The result can optionally be returned as a 
matrix based on the \code{as_matrix} argument.
}
\examples{
A <- matrix(1:15, ncol = 3, byrow = TRUE)
print(A)
stm_keep_rows(A, c(1, 3))

}
\seealso{
\code{\link{stm_keep_cols}}, \code{\link{stm_delete_rows}}, \code{\link{stm_keep_cols}}
}
