% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stm_constants.R
\docType{data}
\name{stm_levy_beta}
\alias{stm_levy_beta}
\title{Levy beta constant}
\format{
A numeric value.
}
\usage{
stm_levy_beta
}
\description{
A predefined constant for the Levy beta constant \code{pi^2 / (12 * log(2))}.
}
\keyword{datasets}
