% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stm_constants.R
\docType{data}
\name{stm_levy_ebeta}
\alias{stm_levy_ebeta}
\title{Exponential of Levy beta constant}
\format{
A numeric value.
}
\usage{
stm_levy_ebeta
}
\description{
A predefined constant for the exponential of the Levy beta constant: \code{exp(pi^2 / (12 * log(2)))}.
}
\keyword{datasets}
