% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stm_linlog_range.R
\name{stm_linspace}
\alias{stm_linspace}
\title{Generate linearly spaced vector}
\usage{
stm_linspace(from, to, n = 50, digits = NULL)
}
\arguments{
\item{from}{A numeric value representing the start of the sequence.}

\item{to}{A numeric value representing the end of the sequence.}

\item{n}{Optional. A positive integer specifying the number of points in 
the sequence. Default is 50.}

\item{digits}{Optional. An integer specifying the number of decimal places 
to round the result. Default is \code{NULL}.}
}
\value{
A numeric vector of length \code{n} with linearly spaced values 
        from \code{from} to \code{to}.
}
\description{
\code{stm_linspace()} generates a vector of linearly spaced values from 
\code{from} to \code{to}, with a specified number of points \code{n}.
}
\details{
This function generates a sequence of values with equal spacing 
between them, which is useful for creating plots or numerical grids.
If \code{digits} is provided, the resulting vector will be rounded to 
the specified number of decimal places.
}
\examples{
stm_linspace(from = 0, to = 1, n = 7)
stm_linspace(from = 0, to = 1, n = 7, digits = 4)

}
\seealso{
\code{\link{stm_logspace}}, \code{\link{stm_arange}}, \code{\link{seq}}, \code{\link{seq_along}}, \code{\link{seq_len}}
}
