% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stm_constants.R
\docType{data}
\name{stm_ln2}
\alias{stm_ln2}
\title{Natural logarithm of 2}
\format{
A numeric value.
}
\usage{
stm_ln2
}
\description{
A predefined constant for the natural logarithm of 2: \code{log(2)}.
}
\keyword{datasets}
