% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stm_constants.R
\docType{data}
\name{stm_mu}
\alias{stm_mu}
\title{Connective constant}
\format{
A numeric value.
}
\usage{
stm_mu
}
\description{
A predefined constant for the connective constant: \code{sqrt(2 + sqrt(2))}.
}
\keyword{datasets}
