% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stm_constants.R
\docType{data}
\name{stm_omega}
\alias{stm_omega}
\title{Omega constant}
\format{
A numeric value.
}
\usage{
stm_omega
}
\description{
A predefined constant for the omega constant: \code{0.56714329040978387299}.
}
\keyword{datasets}
