% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stm_ones.R
\name{stm_ones}
\alias{stm_ones}
\title{Create a matrix of ones}
\usage{
stm_ones(nrows, ncols = nrows)
}
\arguments{
\item{nrows}{The number of rows in the matrix.}

\item{ncols}{The number of columns in the matrix. Defaults to \code{nrows}.}
}
\value{
A matrix filled with ones, with dimensions specified by 
        \code{nrows} and \code{ncols}.
}
\description{
\code{stm_ones()} creates a matrix of ones with the specified number of
rows and columns. It uses the \code{stm_constant()} function to fill
the matrix with ones.
}
\examples{
stm_ones(nrows = 4)
stm_ones(nrows = 3, ncols = 3)
stm_ones(nrows = 2, ncols = 4)

}
\seealso{
\code{\link{stm_ones_like}}, \code{\link{stm_zeros}}, \code{\link{stm_zeros_like}}, \code{\link{stm_constant}}, \code{\link{stm_constant_like}}
}
