% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stm_arrays_like.R
\name{stm_ones_like}
\alias{stm_ones_like}
\title{Generate a matrix or vector of ones like an existing matrix or vector}
\usage{
stm_ones_like(vmat)
}
\arguments{
\item{vmat}{A matrix or vector whose dimensions will be used to generate
the matrix or vector of ones.}
}
\value{
A matrix or vector of the same dimensions as \code{vmat}, filled
        with ones.
}
\description{
\code{stm_ones_like()} generates a matrix or vector of ones that matches
the dimensions of the provided input matrix or vector.
}
\details{
The function checks the validity of the input \code{vmat}, converting it
to a matrix if it is a vector. It then generates a matrix of ones of the
same size as \code{vmat}.
}
\examples{
A <- matrix(1:12, nrow = 3, byrow = TRUE)
stm_ones_like(A)
stm_ones_like(c(1, 2, 3))

}
\seealso{
\code{\link{stm_ones}}
}
