% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stm_pascal.R
\name{stm_pascal}
\alias{stm_pascal}
\title{Generate Pascal's Triangle}
\usage{
stm_pascal(n)
}
\arguments{
\item{n}{The number of rows to generate in Pascal's Triangle. Must be a 
positive integer.}
}
\value{
A list containing:
  \item{y}{A matrix representing Pascal's Triangle.}
  \item{coefs}{A vector of the coefficients in the last row of the triangle.}
}
\description{
\code{stm_pascal()} generates Pascal's Triangle up to the \code{n}-th row
and returns the triangle matrix along with the coefficients in the 
last row.
}
\examples{
stm_pascal(3)
stm_pascal(5)

}
