% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stm_constants.R
\docType{data}
\name{stm_phi}
\alias{stm_phi}
\title{Golden ratio}
\format{
A numeric value.
}
\usage{
stm_phi
}
\description{
A predefined constant for the golden ratio: \code{(1 + sqrt(5)) / 2}.
}
\keyword{datasets}
