% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stm_constants.R
\docType{data}
\name{stm_pi}
\alias{stm_pi}
\title{Mathematical constant pi}
\format{
A numeric value.
}
\usage{
stm_pi
}
\description{
A predefined constant for the mathematical constant pi: \code{pi}.
}
\keyword{datasets}
