% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stm_constants.R
\docType{data}
\name{stm_pi_ln2}
\alias{stm_pi_ln2}
\title{Pi divided by ln(2)}
\format{
A numeric value.
}
\usage{
stm_pi_ln2
}
\description{
A predefined constant for pi divided by the natural logarithm of 2: \code{pi / log(2)}.
}
\keyword{datasets}
