% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stm_constants.R
\docType{data}
\name{stm_pi_log2}
\alias{stm_pi_log2}
\title{Pi divided by log2}
\format{
A numeric value.
}
\usage{
stm_pi_log2
}
\description{
A predefined constant for pi divided by the base-2 logarithm of 2: \code{pi / log(2)}.
}
\keyword{datasets}
