% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stm_trigs.R
\name{stm_rad_to_degrees}
\alias{stm_rad_to_degrees}
\title{Convert radians to degrees}
\usage{
stm_rad_to_degrees(radians, digits = NULL)
}
\arguments{
\item{radians}{A numeric value representing the angle in radians.}

\item{digits}{Optional. The number of decimal places to round the result. 
Default is \code{NULL} for no rounding.}
}
\value{
The value in degrees, rounded to the specified number of decimal 
        places if \code{digits} is provided.
}
\description{
\code{stm_rad_to_degrees()} converts a value in radians to degrees. The 
result can be rounded to a specified number of decimal places.
}
\details{
The function converts the given angle in radians to degrees by multiplying 
by \code{180 / pi}. If \code{digits} is specified, the result is rounded 
to that number of decimal places.
}
\examples{
stm_rad_to_degrees(pi)
stm_rad_to_degrees(c(pi/2, pi, 2 * pi))
stm_rad_to_degrees(pi/180)

}
\seealso{
\code{\link{stm_deg_to_radians}}
}
