% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stm_random_matrix.R
\name{stm_random_matrix}
\alias{stm_random_matrix}
\title{Generate a random matrix}
\usage{
stm_random_matrix(
  min = -4,
  max = 9,
  nrows = 1,
  ncols = nrows,
  replace = TRUE,
  set_seed = NULL,
  row_names = NULL,
  col_names = NULL
)
}
\arguments{
\item{min}{Minimum value in the matrix. Default is \code{-4}.}

\item{max}{Maximum value in the matrix. Default is \code{9}.}

\item{nrows}{Number of rows in the matrix. Default is \code{1}.}

\item{ncols}{Number of columns in the matrix. Default is equal to \code{nrows}.}

\item{replace}{Logical value indicating whether sampling should be with 
replacement. Default is \code{TRUE}.}

\item{set_seed}{Optional seed value for random number generation.}

\item{row_names}{Optional character vector of row names.}

\item{col_names}{Optional character vector of column names.}
}
\value{
A matrix of random numbers with the specified size, range, and names.
}
\description{
\code{stm_random_matrix()} generates a matrix of random numbers with 
specified size and range. Optionally, it allows the user to set the seed 
for reproducibility and assign row and column names.
}
\details{
The function generates a matrix of random numbers within the specified 
range (\code{min} to \code{max}) and with the specified dimensions 
(\code{nrows} and \code{ncols}). If \code{replace} is \code{TRUE}, sampling 
is done with replacement. If \code{set_seed} is provided, the random number 
generation is reproducible. Additionally, row and column names can be specified. 
If the names don't match the matrix dimensions, an error is raised.
}
\examples{
# note that your values will be different
stm_random_matrix()
stm_random_matrix(min = 0, max = 1, nrows = 3, ncols = 3)
stm_random_matrix(min = 10, max = 99, nrows = 5, ncols = 4)

}
\seealso{
\code{\link{stm_random_char}}, \code{\link{stm_random_number}}, \code{\link{stm_random_vector}}, \code{\link{stm_random_matrix}}
}
