% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stm_random_number.R
\name{stm_random_number}
\alias{stm_random_number}
\title{Generate a random number from a vector}
\usage{
stm_random_number(vec = 10:99, set_seed = NULL)
}
\arguments{
\item{vec}{A vector from which to sample. Default is \code{10:99}.}

\item{set_seed}{Optional seed value for random number generation. Default 
is \code{NULL}}
}
\value{
A random number sampled from the specified vector.
}
\description{
\code{stm_random_number()} generates a random number from a specified 
vector. Optionally, the user can set the seed for reproducibility.
}
\details{
The function samples one random number from the provided vector \code{vec}. 
If \code{set_seed} is specified, the random number generation is 
reproducible. The function returns a single sampled value.
}
\examples{
stm_random_number()
stm_random_number(vec = c(1, 2, 3, 4, 5))
stm_random_number(set_seed = 42)

}
\seealso{
\code{\link{stm_random_char}}, \code{\link{stm_random_vector}}, \code{\link{stm_random_matrix}}
}
