% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stm_random_vector.R
\name{stm_random_vector}
\alias{stm_random_vector}
\title{Generate a random vector}
\usage{
stm_random_vector(
  min = 10,
  max = 99,
  n = 10,
  replace = TRUE,
  set_seed = NULL,
  is_sorted = FALSE,
  is_decreasing = FALSE
)
}
\arguments{
\item{min}{The minimum value of the vector elements. Default is 10.}

\item{max}{The maximum value of the vector elements. Default is 99.}

\item{n}{The number of elements in the vector. Default is 10.}

\item{replace}{A logical indicating whether sampling is with replacement.
Default is \code{TRUE}.}

\item{set_seed}{Optional seed for random number generation.}

\item{is_sorted}{A logical indicating whether to sort the vector. 
Default is \code{FALSE}.}

\item{is_decreasing}{A logical indicating whether to sort the vector 
in decreasing order. Default is \code{FALSE}.}
}
\value{
A numeric vector of random values, optionally sorted.
}
\description{
\code{stm_random_vector()} generates a random vector with values between 
the specified minimum and maximum. Optionally, the vector can be sorted 
in increasing or decreasing order.
}
\details{
This function generates a random vector with the specified number of 
elements and value range. The vector can be sorted in increasing or 
decreasing order based on the \code{is_sorted} and \code{is_decreasing} 
parameters.
}
\examples{
stm_random_vector()
stm_random_vector(n = 5, is_sorted = TRUE)
stm_random_vector(
  n = 8, is_sorted = TRUE, is_decreasing = TRUE, set_seed = 42
)

}
\seealso{
\code{\link{stm_random_char}}, \code{\link{stm_random_number}}, \code{\link{stm_random_matrix}}
}
