% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stm_repeat.R
\name{stm_repeat}
\alias{stm_repeat}
\title{Repeat elements of a vector or matrix}
\usage{
stm_repeat(dat, times, is_each = FALSE, on_columns = TRUE)
}
\arguments{
\item{dat}{A vector, matrix, or scalar to be repeated.}

\item{times}{A numeric vector or scalar indicating the number of times 
elements should be repeated.}

\item{is_each}{Logical. If \code{TRUE}, each element of \code{dat} will 
be repeated individually. Default is \code{FALSE}.}

\item{on_columns}{Logical. If \code{TRUE}, the repetition is done 
column-wise for matrices. Default is \code{TRUE}.}
}
\value{
A vector or matrix with repeated elements.
}
\description{
\code{stm_repeat()} repeats elements of a vector or matrix. It can repeat 
based on the entire vector or column-wise for matrices. Optionally, 
it can repeat each element individually.
}
\details{
This function repeats the elements of a vector or matrix. For vectors, 
elements can be repeated individually. For matrices, repetition can 
be done either row-wise or column-wise.
}
\examples{
v = c(4, 5, 8, 3)
stm_repeat(v, times = 3)
stm_repeat(v, times = c(3, 2, 1, 5))
A <- matrix(1:6, nrow = 2, byrow = TRUE)
print(A)
stm_repeat(A, times = 3)
stm_repeat(A, times = 3, on_columns = FALSE)

}
