% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stm_rosser.R
\name{stm_rosser}
\alias{stm_rosser}
\title{Generate the Rosser matrix}
\usage{
stm_rosser()
}
\value{
An 8x8 matrix of predefined values.
}
\description{
\code{stm_rosser()} generates a specific 8x8 matrix used in numerical 
examples. The matrix elements are predefined and arranged in a specific 
pattern.
}
\details{
The function returns a matrix that is commonly used in numerical linear 
algebra problems, known as the Rosser matrix.
}
\examples{
stm_rosser()

}
