% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stm_tranform_fliprotate.R
\name{stm_rot90}
\alias{stm_rot90}
\title{Rotate a matrix by 90 degrees}
\usage{
stm_rot90(mat, k = 1)
}
\arguments{
\item{mat}{A numeric matrix to be rotated.}

\item{k}{The number of 90-degree rotations. Default is \code{1}, which 
rotates the matrix once. The matrix can be rotated multiple 
times by specifying higher values for \code{k}.}
}
\value{
A rotated matrix.
}
\description{
\code{stm_rot90()} rotates a matrix by 90 degrees counterclockwise. The 
number of 90-degree rotations is determined by the argument \code{k}.
}
\details{
The function supports multiple rotations by accepting an integer 
         \code{k}. The matrix is rotated counterclockwise by 90 degrees 
         for each rotation. The parameter \code{k} is modulo 4 to handle 
         the cases where more than 4 rotations are requested.
}
\examples{
A <- matrix(1:12, nrow = 3, byrow = TRUE)
print(A)
stm_rot90(A)
stm_rot90(A, k = 2)
stm_rot90(A, k = -2)
stm_rot90(A, k = 3)
stm_rot90(A, k = 4) # original matrix

}
