% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stm_constants.R
\docType{data}
\name{stm_second_hermite_gamma}
\alias{stm_second_hermite_gamma}
\title{Second Hermite constant}
\format{
A numeric value.
}
\usage{
stm_second_hermite_gamma
}
\description{
A predefined constant for the second Hermite constant: \code{2 / sqrt(3)}.
}
\keyword{datasets}
