% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stm_constants.R
\docType{data}
\name{stm_siver_ratio_delta}
\alias{stm_siver_ratio_delta}
\title{Silver ratio}
\format{
A numeric value.
}
\usage{
stm_siver_ratio_delta
}
\description{
A predefined constant for the silver ratio: \code{sqrt(2) + 1}.
}
\keyword{datasets}
