% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stm_constants.R
\docType{data}
\name{stm_sqrt2}
\alias{stm_sqrt2}
\title{Square root of 2}
\format{
A numeric value.
}
\usage{
stm_sqrt2
}
\description{
A predefined constant for the square root of 2: \code{sqrt(2)}.
}
\keyword{datasets}
