% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stm_constants.R
\docType{data}
\name{stm_sqrt3}
\alias{stm_sqrt3}
\title{Square root of 3}
\format{
A numeric value.
}
\usage{
stm_sqrt3
}
\description{
A predefined constant for the square root of 3: \code{sqrt(3)}.
}
\keyword{datasets}
