% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stm_constants.R
\docType{data}
\name{stm_sqrt5}
\alias{stm_sqrt5}
\title{Square root of 5}
\format{
A numeric value.
}
\usage{
stm_sqrt5
}
\description{
A predefined constant for the square root of 5: \code{sqrt(5)}.
}
\keyword{datasets}
