% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stm_tranform_stack.R
\name{stm_stack_cols}
\alias{stm_stack_cols}
\title{Stack columns together into a data frame or matrix}
\usage{
stm_stack_cols(...)
}
\arguments{
\item{...}{Data frames, matrices, or vectors to be combined as columns.}
}
\value{
A data frame or matrix resulting from column-wise binding of the 
        input data.
}
\description{
\code{stm_stack_cols()} combines multiple data frames or vectors by
column binding them together using \code{cbind()}.
}
\details{
This function stacks the provided data frames, matrices, or vectors as 
columns using the \code{cbind()} function. The resulting object is returned
as a data frame or matrix.
}
\examples{
A <- matrix(1:4, nrow = 2, byrow = TRUE)
print(A)
B <- matrix(5:8, nrow = 2, byrow = TRUE)
print(B)
stm_stack_cols(A, B)

}
\seealso{
\code{\link{cbind}}, \code{\link{rbind}}, \code{\link{stm_stack_rows}}
}
