% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stm_tri_lower_upper.R
\name{stm_triu}
\alias{stm_triu}
\title{Extract upper triangular part of a matrix}
\usage{
stm_triu(mat, k = 0)
}
\arguments{
\item{mat}{A numeric matrix.}

\item{k}{An integer indicating the diagonal to be used. Default is \code{0},
which returns the upper triangular part including the main diagonal.}
}
\value{
A matrix containing the upper triangular part of \code{mat}, 
        with the specified diagonal offset.
}
\description{
\code{stm_triu()} returns the upper triangular part of a matrix, optionally 
with a specified diagonal offset.
}
\details{
This function calls \code{stm_tri_lower_upper()} to extract the upper 
triangular part of the matrix.
}
\examples{
A <- matrix(1:16, nrow = 4, byrow = TRUE)
print(A)
stm_triu(A)

}
\seealso{
\code{\link{stm_tril}}
}
