% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stm_matrix_lus.R
\name{stm_triu_from_vector}
\alias{stm_triu_from_vector}
\title{Create an upper triangular matrix from a vector}
\usage{
stm_triu_from_vector(vec)
}
\arguments{
\item{vec}{A numeric vector. The vector from which the upper triangular
matrix is created.}
}
\value{
An upper triangular matrix created from the input vector.
}
\description{
\code{stm_triu_from_vector()} generates an upper triangular matrix from
a given vector. The function calls another matrix function to generate
the matrix.
}
\details{
This function relies on \code{stm_matrix_lus()} to generate the 
         matrix.
}
\examples{
v <- c(1, 2, 3, 4, 5, 6)
print(v)
stm_triu_from_vector(v)

}
\seealso{
\code{\link{stm_tril_from_vector}}, \code{\link{stm_symmetric_from_vector}}
}
