% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stm_unpack.R
\name{stm_unpack}
\alias{stm_unpack}
\title{Unpack elements of a data structure into the environment}
\usage{
stm_unpack(dat, output = FALSE, envir = .GlobalEnv)
}
\arguments{
\item{dat}{A list, named vector, or data frame to unpack into variables.}

\item{output}{Logical. If \code{TRUE}, displays the unpacked variable names
and their values. Default is \code{FALSE}.}

\item{envir}{The environment in which to unpack variables. Default is
\code{.GlobalEnv}.}
}
\value{
Invisibly returns \code{NULL}. The function creates variables in
        the specified environment.
}
\description{
\code{stm_unpack()} assigns elements of a list, named vector, or data frame
to the specified environment, creating variables for each element.
}
\details{
This function allows you to unpack elements of a supported data structure
(list, named vector, or data frame) and assign them as variables in the
specified environment. If \code{data} is not of a supported type, an error
is raised.
}
\examples{
lst <- list(a = 10, b = 20, c = 30)
print(lst)
stm_unpack(lst)
a
b
c

v <- c(x = 1, y = 2, z = 3)
print(v)
stm_unpack(v)
a
b
c

}
