% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stm_vander.R
\name{stm_vander}
\alias{stm_vander}
\title{Generate a Vandermonde matrix from a vector}
\usage{
stm_vander(vec, ncols = NULL, is_increasing = FALSE)
}
\arguments{
\item{vec}{A numeric vector for which the Vandermonde matrix will be 
created.}

\item{ncols}{Optional. An integer specifying the number of columns in the 
matrix. Default is \code{NULL}, which keeps all columns.}

\item{is_increasing}{A logical value indicating whether the matrix should 
be in increasing order. Default is \code{FALSE}.}
}
\value{
A matrix representing the Vandermonde matrix, either in the default 
        order or reversed if \code{is_increasing = TRUE}.
}
\description{
\code{stm_vander()} creates a Vandermonde matrix from a given vector. It 
raises each element of the vector to the power of integers from 0 to 
\code{n-1}, where \code{n} is the length of the vector. The user can 
specify the number of columns in the matrix and whether the matrix should 
be in increasing order.
}
\details{
The function generates a Vandermonde matrix using the outer product, where 
the elements of \code{vec} are raised to successive powers from 0 to 
\code{n-1}. The user can specify the number of columns using the
\code{ncols} parameter. The matrix can also be returned in increasing order
if \code{is_increasing} is set to \code{TRUE}.
}
\examples{
v <- c(1, 2, 3, 4)
print(v)
stm_vander(v)
stm_vander(v, is_increasing = TRUE)

}
