% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stm_zeros.R
\name{stm_zeros}
\alias{stm_zeros}
\title{Generate a zero matrix}
\usage{
stm_zeros(nrows, ncols = nrows)
}
\arguments{
\item{nrows}{Optional. The number of rows in the matrix. Defaults to
\code{1} if not specified. Must be a positive integer.}

\item{ncols}{Optional. The number of columns in the matrix. If \code{NULL},
it defaults to the same value as \code{nrows}. Must also be
a positive integer.}
}
\value{
A matrix filled with zeros, with dimensions defined by \code{nrows}
        and \code{ncols}. Returns an empty matrix with dimensions
        \code{(0, 0)} if either dimension is zero.
}
\description{
Returns a matrix filled with zeros, with dimensions specified by the user.
If dimensions are not provided, a 1 x 1 zero matrix will be returned.
}
\examples{
stm_zeros(nrows = 4)
stm_zeros(nrows = 3, ncols = 3)
stm_zeros(nrows = 2, ncols = 4)

}
