% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stm_arrays_like.R
\name{stm_zeros_like}
\alias{stm_zeros_like}
\title{Generate a matrix or vector of zeros like an existing matrix or vector}
\usage{
stm_zeros_like(vmat)
}
\arguments{
\item{vmat}{A matrix or vector whose dimensions will be used to generate
the matrix or vector of zeros.}
}
\value{
A matrix or vector of the same dimensions as \code{vmat}, filled
        with zeros.
}
\description{
\code{stm_zeros_like()} generates a matrix or vector of zeros that matches
the dimensions of the provided input matrix or vector.
}
\details{
The function checks the validity of the input \code{vmat}, converting it
to a matrix if it is a vector. It then generates a matrix of zeros of the
same size as \code{vmat}.
}
\examples{
A <- matrix(1:12, nrow = 3, byrow = TRUE)
stm_zeros_like(A)
stm_zeros_like(c(1, 2, 3))

}
\seealso{
\code{\link{stm_zeros}}
}
