% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stm_transform_swap.R
\name{stm_swap_vector}
\alias{stm_swap_vector}
\alias{stm_swap_rows}
\alias{stm_swap_cols}
\title{Swap elements in a vector, rows, or columns}
\usage{
stm_swap_vector(dat, i, j)

stm_swap_rows(dat, i, j)

stm_swap_cols(dat, i, j)
}
\arguments{
\item{dat}{The input vector, matrix, or data frame. In
\code{stm_swap_vector}, it should be a vector. In
\code{stm_swap_rows} and \code{stm_swap_cols}, it should be a matrix or 
data frame.}

\item{i}{The index of the first element (or row/column) to swap.}

\item{j}{The index of the second element (or row/column) to swap.}
}
\value{
A vector, matrix, or data frame with the specified elements swapped.
}
\description{
The \code{stm_swap...()} set of functions swap the elements at specified 
indices in a vector, rows in a matrix or data frame, or columns in a matrix 
or data frame.
}
\examples{
# vector
v <- c(1, 2, 3, 4, 5)
print(v)
stm_swap_vector(v, i = 2, j = 4)

# swap matrix rows
A <- matrix(1:9, ncol = 3, byrow = TRUE)
print(A)
stm_swap_rows(A, 1, 3)

# swap matrix columns
A <- matrix(1:9, ncol = 3, byrow = TRUE)
print(A)
stm_swap_cols(A, 1, 3)

}
